+++ src/messaging/message.asserts.ts
@@ -13,8 +13,9 @@ import {
 } from '../model/internal-messages.model'
 import { Dapp, PodActions, PodPermission } from '../model/storage/dapps.model'
 import { Network } from '../model/storage/network.model'
-import { KeyData, StorageSession } from '../model/storage/session.model'
+import { Session } from '../model/storage/session.model'
 import { Swarm } from '../model/storage/swarm.model'
+import { BytesMessage } from './scripts.messaging'
 
 export function isString(data: unknown): data is string {
   return typeof data === 'string'
@@ -82,19 +83,14 @@ export function isSwarm(data: unknown): data is Swarm {
   return typeof extensionId === 'string'
 }
 
-export function isStorageKeyData(data: unknown): data is KeyData<string> {
-  const { seed, url } = (data || {}) as KeyData<string>
-
-  return typeof seed === 'string' && typeof url === 'string'
-}
-
-export function isStorageSession(data: unknown): data is StorageSession {
-  const { ensUserName, localUserName, network, key } = (data || {}) as StorageSession
+export function isSession(data: unknown): data is Session {
+  const { ensUserName, localUserName, network, seed, address } = (data || {}) as Session
 
   return (
     (typeof ensUserName === 'string' || typeof localUserName === 'string') &&
     isNetwork(network) &&
-    isStorageKeyData(key)
+    isAddress(address) &&
+    seed?.byteLength === 64
   )
 }
 
@@ -137,4 +133,10 @@ export function assertBeeUrl(url: string): asserts url {
   if (!url || !(url.startsWith('http://') || url.startsWith('https://'))) {
     throw new Error('Blossom: Invalid Bee URL')
   }
+}
+
+export function isSerializedUint8Array(data: unknown): data is BytesMessage {
+  const { type, value } = (data || {}) as BytesMessage
+
+  return type === 'bytes' && isString(value)
 }