+++ src/model/error.model.ts
@@ -1,0 +1,7 @@
+import { ErrorCode } from '../constants/errors'
+
+export interface ErrorObject {
+  code: ErrorCode
+  priority: number
+  message: string
+}